+++ src/index.tsx
@@ -16,17 +16,17 @@ const rootElement = document.getElementById('root');
 if (!rootElement) throw new Error('Failed to find the root element');
 const root = ReactDOM.createRoot(rootElement);
 root.render(
-  // <React.StrictMode>
-  <QueryClientProvider client={queryClient}>
-    <RecoilRoot>
-      <Suspense>
-        <ThemeProvider theme={defaultTheme}>
-          <App />
-        </ThemeProvider>
-      </Suspense>
-    </RecoilRoot>
-  </QueryClientProvider>
-  // </React.StrictMode>
+  <React.StrictMode>
+    <QueryClientProvider client={queryClient}>
+      <RecoilRoot>
+        <Suspense>
+          <ThemeProvider theme={defaultTheme}>
+            <App />
+          </ThemeProvider>
+        </Suspense>
+      </RecoilRoot>
+    </QueryClientProvider>
+  </React.StrictMode>
 );
 
 reportWebVitals();
