+++ src/pages/WelcomePage.tsx
@@ -1,7 +1,8 @@
-import React, { useEffect, useInsertionEffect } from 'react';
+import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import styled from 'styled-components';
+
 import WelcomePic from '../components/common/elem/WelcomePic';
 
 import { userProfile } from '../recoil/userAtoms';
@@ -17,36 +18,29 @@ const WelcomePage = () => {
 
   return (
     <Wrapper>
-      <ContentWrapper>
-        <Text>
-          {name} 님 <br />
-          가입을 환영합니다!
-        </Text>
-
-        <WelcomePic />
-      </ContentWrapper>
+      <Text>
+        {name} 님 <br />
+        가입을 환영합니다!
+      </Text>
+      <WelcomePic />
     </Wrapper>
   );
 };
 
 const Wrapper = styled.div`
+  position: absolute;
   display: flex;
   flex-direction: column;
+  align-items: center;
   width: 100%;
   height: 100%;
-`;
-
-const ContentWrapper = styled(Wrapper)`
-  align-items: center;
-  justify-content: center;
-  width: 100%;
   gap: 10px;
+  top: 200px;
 `;
 
 const Text = styled.div`
   font: ${(props) => props.theme.headingH2};
   text-align: center;
-  margin: 20px;
 `;
 
 export default WelcomePage;
